requirements.txt
#semantic-kernel

#langfuse>=3.6.1,<4.0.0

#openlit>=1.35.5,<2.0.0

#chainlit>=2.8.3,<3.0.0

#agent-framework==1.0.0b251007

#python-dotenv>=1.1.1,<2.0.0

chainlit

python-dotenv

openai

#kubernetes

azure-mgmt-storage

azure-identity

azure-mgmt-resource

azure-mgmt-network

azure-mgmt-compute

#langfuse

#tiktoken

#httpx
